Split the API tests into "happy path"/"errors"
The single test file is starting to get large and unwieldy, and I'm
about to pile even more tests into it -- let's break it up to make it
more manageable.
